SITCAD/backend/requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
cloud-sql-python-connector[pg8000]
pg8000
firebase_admin
Pillow